refactored into bitcoin.ts
